connection manager instantiates in layer1-chain
